reseption/requirements.txt
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
Django==5.2.2
django-environ==0.12.0
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-translate==3.20.2
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==5.4.0
outcome==1.3.0.post0
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.4
rfc3986==1.5.0
rsa==4.9.1
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
